Execution control
This change contains changes to the behaviour of the
framework especially on x86 and rpi hosts. The
changes aim to make the frameworks operation as
similar as possible in both cases (the rpi
/boot/wirepas entrypoint remains the key difference).

The change introduced within with commit allow
one to select which features should be enable
on particular hosts. Closes #11.

Also, most if not all environment keys have been
moved to files under environment, making it
much easier to update and track default values.