Docker services setup
This change makes away with the eval
call when redeploying the docker services.

The execution prints are also updated and
provide better traceability on which
image is running (and if an update
happened due to it).

The transport and sink service also receive
further exposure to the custom.env file,
allowing the user to define the restart
policy and container user.

The logging can also be routed through fluentd
to retrieve gateway telemetry data.

Relates to wirepas/gateway#79