Network management update
This change ensure that the WiFI interface is
up and that the avahi file is not copied if
it is already present in the system.

Public keys are also appended to the authorized
keys if the user is also set. No need to have
a password set for such purpose.